+++ stdpp/gmultiset.v
@@ -332,8 +332,16 @@ Ltac multiset_simplify_singletons :=
   end.
 
 (** Putting it all together *)
-Ltac multiset_solver :=
-  set_solver by (multiset_instantiate; multiset_simplify_singletons; lia).
+(** Similar to [set_solver] and [naive_solver], [multiset_solver] has a [by]
+parameter whose default is [eauto]. *)
+Tactic Notation "multiset_solver" "by" tactic3(tac) :=
+  set_solver by (multiset_instantiate;
+                 multiset_simplify_singletons;
+                 (* [fast_done] to solve trivial equalities or contradictions,
+                 [lia] for the common case that involves arithmetic,
+                 [tac] if all else fails *)
+                 solve [fast_done|lia|tac]).
+Tactic Notation "multiset_solver" := multiset_solver by eauto.
 
 Section more_lemmas.
   Context `{Countable A}.
@@ -659,13 +667,7 @@ Section more_lemmas.
   Proof. multiset_solver. Qed.
   Lemma gmultiset_singleton_subseteq x y :
     {[+ x +]} ⊆@{gmultiset A} {[+ y +]} ↔ x = y.
-  Proof.
-    split; [|multiset_solver].
-    (* FIXME: [multiset_solver] should solve this *)
-    intros Hxy. specialize (Hxy x).
-    rewrite multiplicity_singleton, multiplicity_singleton' in Hxy.
-    case_decide; [done|lia].
-  Qed.
+  Proof. multiset_solver. Qed.
 
   Lemma gmultiset_elem_of_subseteq X1 X2 x : x ∈ X1 → X1 ⊆ X2 → x ∈ X2.
   Proof. multiset_solver. Qed.